fix(search): delegate scroll-to-result from DocSearchOverlay to MarkdownPreview
Scroll after semantic search was silently failing because the overlay
tried to manipulate the DOM after closing itself. The fix delegates all
scroll logic: DocSearchOverlay now accepts an onscrollto callback; the
page computes the target block index from the chunk_index and calls
previewRef.scrollToBlock(), which uses the same scrollIntoView path as
the working scrollToComment method.
